+++ skeleton_project/source/driver/queueLogic.c
@@ -100,3 +100,12 @@ void update_queue_internal(Queue* queue){
     }
 
 }
+
+
+void input_floor(Queue* queue){
+    for(int floor = 0; floor < N_FLOORS; floor++){
+        for(int button = 0; button < N_BUTTONS; button++){
+            printf("placeholder");
+        }
+    }
+}
