+++ skeleton_project/source/driver/queueLogic.h
@@ -14,12 +14,7 @@ typedef struct{
     QueueDirection direction;
 }Queue; 
 
-extern struct Queue G_QUEUE;
-
 void init_queue(Queue*);
-
-
-
 void update_queue_internal(Queue*);
 void set_queue_last(Queue*, int);
 void delete_queue_element(Queue*, int);
@@ -29,3 +24,5 @@ int set_queue_next(Queue*);
 int sum_current_queue(Queue*);
 int get_queue_last();
 int get_queue_next();
+
+void input_floor(Queue*);
