--- skeleton_project/source/driver/queue_input.c
@@ -1,19 +1,0 @@
-#include <assert.h>
-#include <stdlib.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netdb.h>
-#include <stdio.h>
-#include <pthread.h>
-
-#include "elevio.h"
-#include "queue_input.h"
-
-
-void input_floor(Queue* queue){
-    for(int floor = 0; floor < N_FLOORS; floor++){
-        for(int button = 0; button < N_BUTTONS; button++){
-            printf("placeholder");
-        }
-    }
-}
