--- skeleton_project/source/driver/queue_input.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void input_floor(Queue*);
